### 2025-04-08 #8029 @editorialbot
[PRE REVIEW]: HySetter: A Python Package for Reproducible Hydroclimate Data Subsetting over CONUS
**Submitting author:** @cheginit (<a href="http://orcid.org/0000-0002-5430-6000">Taher Chegini</a>)
**Repository:** https://github.com/hyriver/hysetter
**Branch with paper.md** (empty if default branch):
**Version:** v0.3.0
