### 2025-04-25 #8029 @kthyng
@cheginit It'd be great if you can shorten your paper by a few hundred words — maybe move a section to the docs?
