### 2025-04-29 2379f47
docs: reduce the word count. [skip ci]
