+++ docs/paper/paper.md
@@ -21,67 +21,50 @@ bibliography: paper.bib
 
 # Summary
 
-Hydrological modeling and research increasingly requires integrating diverse
-hydroclimate datasets from various sources, which creates a significant technical burden
-for researchers. This data acquisition process frequently becomes a major bottleneck,
-diverting valuable time and resources away from scientific analysis. HySetter addresses
-this challenge by providing a command-line interface for automated and reproducible
-subsetting of hydroclimate data across the conterminous United States (CONUS). Built
-upon the HyRiver software stack, HySetter streamlines data acquisition through a unified
-configuration file that eliminates the need for writing complex scripts for different
-data sources. By centralizing the data acquisition workflow, HySetter enables
-researchers to focus on scientific questions rather than data preparation, while
-simultaneously enhancing the reproducibility and transparency of hydrological research.
+Hydrological modeling requires integrating diverse hydroclimate datasets, which often
+creates a significant technical burden for researchers. HySetter addresses this
+challenge by providing a command-line interface for automated subsetting of hydroclimate
+data across the conterminous United States (CONUS). Built upon the HyRiver software
+stack, HySetter streamlines data acquisition through a unified configuration file that
+eliminates writing complex scripts for different data sources. By centralizing the
+workflow, HySetter enables researchers to focus on scientific questions rather than data
+preparation, while enhancing the reproducibility of hydrological research.
 
 # Statement of Need
 
-Hydrological research increasingly demands integration of diverse datasets spanning
-multiple spatial and temporal scales. However, the acquisition, processing, and
-integration of these datasets presents several challenges:
+Hydrological research demands integration of diverse datasets spanning multiple scales.
+The acquisition and processing of these datasets presents several challenges: (1)
+Traditional approaches require complex scripts for different data sources, leading to
+fragmented workflows; (2) Researchers often spend more time on data collection than
+scientific analysis; (3) Technical overhead impacts novice researchers with limited
+programming skills; (4) Experienced researchers develop redundant code for data
+acquisition across projects.
 
-1. Traditional approaches require developing complex scripts for different data sources,
-    leading to fragmented workflows that hinder reproducibility.
-1. Researchers often spend more time on data collection than on scientific analysis and
-    interpretation.
-1. The technical overhead disproportionately impacts novice researchers with limited
-    programming skills, hampering their ability to conduct hydrological studies.
-1. Even experienced researchers frequently develop redundant boilerplate code for data
-    acquisition across different projects.
+These challenges undermine the adoption of FAIR principles [@wilkinson2016fair] in
+hydrological research. While tools like HydroMT [@eilander2023hydromt] address similar
+challenges for model building, a comparable solution for hydroclimate data acquisition
+for CONUS is lacking.
 
-These challenges collectively undermine the adoption of FAIR (Findable, Accessible,
-Interoperable, Reusable) principles [@wilkinson2016fair] in hydrological research. While
-tools like HydroMT [@eilander2023hydromt] have addressed similar challenges for model
-building and analysis, a comparable solution specifically targeting hydroclimate data
-acquisition for CONUS remains lacking.
-
-HySetter fills this gap by providing a user-friendly framework that requires minimal
-technical expertise while supporting advanced applications. Through a simple YAML
-configuration file and command-line interface, HySetter democratizes access to
-hydroclimate data, enabling researchers of varying technical backgrounds to efficiently
-obtain consistent, well-documented datasets for their studies.
+HySetter fills this gap by providing a user-friendly framework requiring minimal
+technical expertise while supporting advanced applications. Through a YAML configuration
+file and command-line interface, HySetter democratizes access to hydroclimate data for
[…]
