### 2025-05-24 91d2dab
docs: remove the wrong tag for aligning to center. [skip ci]
