### 2025-05-24 078d497
chore: use any float type for ndarray to fix pyright error. [skip ci]
