### 2025-05-24 ddf9679
docs: add instructions to use pixi instead of micromamba. [skip ci]
