### 2025-05-24 c458162
chore: fix the env name for tests. [skip ci]
