+++ .github/workflows/test.yml
@@ -28,7 +28,7 @@ jobs:
     - uses: actions/checkout@v4
     - uses: prefix-dev/setup-pixi@v0.8.3
       with:
-        pixi-version: v0.42.1
+        pixi-version: v0.47.0
         environments: ${{ matrix.environment }}
         activate-environment: ${{ matrix.environment }}
         cache: true
