### 2025-05-24 2df34e8
chore: rasterio is still not compatible with gdal 3.11. [skip ci]
