+++ environment.yml
@@ -6,6 +6,7 @@ dependencies:
 - python >=3.10
 - hysetter
 # Examples
+- libgdal-core <3.11
 - flox
 - ipykernel
 - ipywidgets
@@ -14,6 +15,5 @@ dependencies:
 - notebook
 - numba
 - numbagg
-- pytest
 - salib
 - sceua
