### 2025-05-24 7c2c76f
chore: bump pixi version. [skip ci]
