+++ .github/workflows/test.yml
@@ -26,7 +26,7 @@ jobs:
 
     steps:
     - uses: actions/checkout@v4
-    - uses: prefix-dev/setup-pixi@v0.8.3
+    - uses: prefix-dev/setup-pixi@v0.8.8
       with:
         pixi-version: v0.47.0
         environments: ${{ matrix.environment }}
