### 2025-05-24 63eeb7b
docs: update to note the use of pixi. [skip ci]
