+++ .github/PULL_REQUEST_TEMPLATE.md
@@ -1,5 +1,5 @@
 <!-- Feel free to remove check-list items aren't relevant to your changes -->
 
 - [ ] Closes #xxxx
-- [ ] Tests added and `nox` passes.
-- [ ] Changes and the contributor name are documented in `HISTORY.rst`.
+- [ ] Tests added and `pixi r -e test310 test` and `pixi r -e test313 test` pass.
+- [ ] Changes and the contributor name are documented in `CHANGELOG.md`.
