### 2025-05-24 a4377ac
docs: update install instructions to use pixi. [skip ci]
