### 2025-05-24 02a6beb
chore: bump the min version for pydaymet. [skip ci]
