+++ pyproject.toml
@@ -42,7 +42,7 @@ dependencies = [
   "planetary-computer",
   "pyarrow>=1.0.1",
   "pydantic>=2.7.1",
-  "pydaymet>=0.19,<0.20",
+  "pydaymet>=0.19.4,<0.20",
   "pygeohydro>=0.19,<0.20",
   "pygridmet>=0.19,<0.20",
   "pynhd>=0.19.4,<0.20",
