### 2025-05-24 9f3b9fc
fix: create the root dir for aoi to address the issue with the overwrite option. [skip ci]
