### 2025-05-24 db04ca4
fix: delete the project folder instead of its root. [skip ci]
