### 2025-05-24 05bec3b
chore: separet the test into two tests so they are shown individually. [skip ci]
