+++ tests/test_hysetter.py
@@ -3,7 +3,7 @@ from __future__ import annotations
 import io
 
 import pytest
-from click.testing import CliRunner
+from click.testing import CliRunner, Result
 
 import hysetter as hs
 from hysetter.cli import cli
@@ -15,9 +15,17 @@ def runner():
     return CliRunner()
 
 
-def test_config(runner: CliRunner) -> None:
+@pytest.fixture
+def run_config_once(runner: CliRunner) -> Result:
     ret = runner.invoke(cli, ["config_demo.yml"])
-    assert ret.exit_code == 0
+    return ret
+
+
+def test_config(run_config_once: Result) -> None:
+    assert run_config_once.exit_code == 0
+
+
+def test_config_overwrite(runner: CliRunner, run_config_once: Result) -> None:
     ret = runner.invoke(cli, ["config_demo.yml", "--overwrite"])
     assert ret.exit_code == 0
 
