### 2025-05-24 29c3a58
docs: update the changelog. [skip ci]
