### 2025-06-19 hyriver/hysetter#6 @mhweber
conda install
### Is your feature request related to a problem?

Perhaps in [Installation section](https://hysetter.readthedocs.io/latest/#installation) show conda install along with pip and micromamba?
@openjournals/joss-reviews#8301

### Describe the solution you'd like

conda install -c conda-forge hysetter

### Describe alternatives you've considered

_No response_

### Additional context

_No response_
