### 2025-06-19 hyriver/hysetter#7 @mhweber
Getting Streamflow from NWIS using config_demo.yml error
### What happened?

Other metrics returned successfully after installing `hysetter` and running `hysetter config_demo.yml` but I get an error on the last data retrieval step of `Getting streamflow from NWIS for 1 stations` with the  lengthy error included below.  I thought it might be related to write permission locally but other data returned and I verified the directory was not read-only.
@openjournals/joss-reviews#8301

### What did you expect to happen?

expected streamflow to return data as other metrics

### Minimal Complete Verifiable Example

```Python
Getting streamflow from NWIS for 1 stations
╭─────────────────────────────── Traceback (most recent call last) ────────────────────────────────╮
│ C:\Users\mwebe\miniforge3\Lib\site-packages\xarray\backends\file_manager.py:211 in               │
│ _acquire_with_cache_info                                                                         │
│                                                                                                  │
│   208 │   │   """Acquire a file, returning the file and whether it was cached."""                │
│   209 │   │   with self._optional_lock(needs_lock):                                              │
│   210 │   │   │   try:                                                                           │
│ ❱ 211 │   │   │   │   file = self._cache[self._key]                                              │
│   212 │   │   │   except KeyError:                                                               │
│   213 │   │   │   │   kwargs = self._kwargs                                                      │
│   214 │   │   │   │   if self._mode is not _DEFAULT_MODE:                                        │
│                                                                                                  │
│ ╭─────────────────────────────────────────── locals ───────────────────────────────────────────╮ │
│ │     kwargs = {                                                                               │ │
│ │              │   'clobber': True,                                                            │ │
│ │              │   'diskless': False,                                                          │ │
│ │              │   'persist': False,                                                           │ │
│ │              │   'format': 'NETCDF4',                                                        │ │
│ │              │   'mode': 'w'                                                                 │ │
│ │              }                                                                               │ │
│ │ needs_lock = True                                                                            │ │
│ │       self = CachingFileManager(<class 'netCDF4._netCDF4.Dataset'>,                          │ │
│ │              'C:\\Users\\mwebe\\OneDrive\\GitProjects\\hysetter\\data\\example_1\\streamflo… │ │
│ │              00:00:00_2018-01-02 00:00:00.nc', mode='w', kwargs={'clobber': True,            │ │
│ │              'diskless': False, 'persist': False, 'format': 'NETCDF4'},                      │ │
│ │              manager_id='d3903d02-af42-49a3-ab03-5bae6912ad42')                              │ │
│ ╰──────────────────────────────────────────────────────────────────────────────────────────────╯ │
│                                                                                                  │
│ C:\Users\mwebe\miniforge3\Lib\site-packages\xarray\backends\lru_cache.py:56 in __getitem__       │
│                                                                                                  │
│    53 │   def __getitem__(self, key: K) -> V:                                                    │
│    54 │   │   # record recent use of the key by moving it to the front of the list               │
│    55 │   │   with self._lock:                                                                   │
│ ❱  56 │   │   │   value = self._cache[key]                                                       │
│    57 │   │   │   self._cache.move_to_end(key)                                                   │
│    58 │   │   │   return value                                                                   │
│    59                                                                                            │
│                                                                                                  │
│ ╭─────────────────────────────────────────── locals ───────────────────────────────────────────╮ │
│ │  key = [                                                                                     │ │
│ │        │   <class 'netCDF4._netCDF4.Dataset'>,                                               │ │
│ │        │   (                                                                                 │ │
│ │        │   │                                                                                 │ │
│ │        'C:\\Users\\mwebe\\OneDrive\\GitProjects\\hysetter\\data\\example_1\\streamflow\\str… │ │
│ │        │   ),                                                                                │ │
│ │        │   'a',                                                                              │ │
│ │        │   (                                                                                 │ │
│ │        │   │   ('clobber', True),                                                            │ │
│ │        │   │   ('diskless', False),                                                          │ │
│ │        │   │   ('format', 'NETCDF4'),                                                        │ │
│ │        │   │   ('persist', False)                                                            │ │
│ │        │   ),                                                                                │ │
│ │        │   'd3903d02-af42-49a3-ab03-5bae6912ad42'                                            │ │
│ │        ]                                                                                     │ │
│ │ self = <xarray.backends.lru_cache.LRUCache object at 0x000002043B4002C0>                     │ │
│ ╰──────────────────────────────────────────────────────────────────────────────────────────────╯ │
╰──────────────────────────────────────────────────────────────────────────────────────────────────╯
KeyError: [<class 'netCDF4._netCDF4.Dataset'>,
('C:\\Users\\mwebe\\OneDrive\\GitProjects\\hysetter\\data\\example_1\\streamflow\\streamflow_2018-01-01 00:00:00_2018-01-02 00:00:00.nc',), 'a',
(('clobber', True), ('diskless', False), ('format', 'NETCDF4'), ('persist', False)), 'd3903d02-af42-49a3-ab03-5bae6912ad42']

╭─────────────────────────────── Traceback (most recent call last) ────────────────────────────────╮
│ C:\Users\mwebe\miniforge3\Lib\site-packages\hysetter\nwis.py:45 in get_streamflow                │
│                                                                                                  │
│   42 │   │   │   │   │   console.print(                                                          │
│   43 │   │   │   │   │   │   f"Failed to get streamflow data for AOI for {data_cfg.use_col} c    │
│   44 │   │   │   │   │   )                                                                       │
│ ❱ 45 │   │   │   │   qobs.to_netcdf(sf_paths[0])                                                 │
│   46 │   │   │   except Exception:                                                               │
│   47 │   │   │   │   console.print_exception(show_locals=True, max_frames=4)                     │
│   48 │   │   │   │   console.print(                                                              │
│                                                                                                  │
│ ╭─────────────────────────────────────────── locals ───────────────────────────────────────────╮ │
│ │      console = <console width=146 ColorSystem.TRUECOLOR>                                     │ │
│ │     data_cfg = Streamflow(                                                                   │ │
│ │                │   start_date=datetime.datetime(2018, 1, 1, 0, 0),                           │ │
│ │                │   end_date=datetime.datetime(2018, 1, 2, 0, 0),                             │ │
│ │                │   frequency='daily',                                                        │ │
│ │                │   within_aoi=False,                                                         │ │
│ │                │   use_col='gage_id'                                                         │ │
│ │                )                                                                             │ │
│ │          end = '2018-01-02'                                                                  │ │
│ │         freq = 'dv'                                                                          │ │
│ │          gdf = │   │   │   │   │   │   │   │   │   │   │   geometry    area  perimeter       │ │
│ │                gage_id                                                                       │ │
│ │                0  POLYGON ((-121.07543 48.37015, -121.07582 48.3...  211500       3000       │ │
│ │                12181200                                                                      │ │
│ │ model_config = Config(                                                                       │ │
│ │                │   project=Project(name='example_1', data_dir='data'),                       │ │
│ │                │   aoi=AOI(                                                                  │ │
│ │                │   │   huc_ids=[],                                                           │ │
│ │                │   │   nhdv2_ids=[],                                                         │ │
│ │                │   │   gagesii_basins=['12181200'],                                          │ │
│ │                │   │   mainstem_main=None,                                                   │ │
│ │                │   │   mainstem_tributaries=None,                                            │ │
│ │                │   │   geometry_file='',                                                     │ │
│ │                │   │   nhdv2_flowlines=True,                                                 │ │
│ │                │   │   streamcat_attrs=['fert', 'bfi'],                                      │ │
│ │                │   │   nldi_attrs=None                                                       │ │
│ │                │   ),                                                                        │ │
│ │                │   forcing=Forcing(                                                          │ │
│ │                │   │   source='daymet',                                                      │ │
│ │                │   │   start_date=datetime.datetime(2018, 1, 1, 0, 0),                       │ │
│ │                │   │   end_date=datetime.datetime(2018, 1, 2, 0, 0),                         │ │
│ │                │   │   variables=['prcp', 'tmin'],                                           │ │
│ │                │   │   crop=False,                                                           │ │
│ │                │   │   geometry_buffer=0.0                                                   │ │
│ │                │   ),                                                                        │ │
│ │                │   topo=Topo(                                                                │ │
│ │                │   │   resolution_m=10,                                                      │ │
│ │                │   │   derived_variables=['slope', 'aspect', 'curvature'],                   │ │
│ │                │   │   crop=True,                                                            │ │
│ │                │   │   geometry_buffer=0.0                                                   │ │
│ │                │   ),                                                                        │ │
│ │                │   soil=Soil(                                                                │ │
│ │                │   │   source='soilgrids',                                                   │ │
│ │                │   │   variables=['bdod_5', 'cec_5'],                                        │ │
│ │                │   │   crop=True,                                                            │ │
│ │                │   │   geometry_buffer=0.0                                                   │ │
│ │                │   ),                                                                        │ │
│ │                │   nlcd=NLCD(                                                                │ │
│ │                │   │   cover=[2016],                                                         │ │
│ │                │   │   impervious=[2016],                                                    │ │
│ │                │   │   canopy=[2016],                                                        │ │
│ │                │   │   descriptor=[2016],                                                    │ │
│ │                │   │   crop=True,                                                            │ │
│ │                │   │   geometry_buffer=0.0                                                   │ │
│ │                │   ),                                                                        │ │
│ │                │   nid=NID(within_aoi=True),                                                 │ │
│ │                │   streamflow=Streamflow(                                                    │ │
│ │                │   │   start_date=datetime.datetime(2018, 1, 1, 0, 0),                       │ │
│ │                │   │   end_date=datetime.datetime(2018, 1, 2, 0, 0),                         │ │
│ │                │   │   frequency='daily',                                                    │ │
│ │                │   │   within_aoi=False,                                                     │ │
│ │                │   │   use_col='gage_id'                                                     │ │
│ │                │   ),                                                                        │ │
│ │                │   remote_rasters=None,                                                      │ │
│ │                │   overwrite=False,                                                          │ │
│ │                │   file_paths=FilePaths(                                                     │ │
│ │                │   │   project=WindowsPath('data/example_1'),                                │ │
│ │                │   │   aoi_parquet=WindowsPath('data/example_1/aoi.parquet'),                │ │
│ │                │   │                                                                         │ │
│ │                flowlines=FileList(C:\Users\mwebe\OneDrive\GitProjects\hysetter\data\example… │ │
│ │                1 files):                                                                     │ │
│ │                C:\Users\mwebe\OneDrive\GitProjects\hysetter\data\example_1\nhdv2_flowlines\… │ │
│ │                │   │                                                                         │ │
│ │                streamcat_attrs=FileList(C:\Users\mwebe\OneDrive\GitProjects\hysetter\data\e… │ │
│ │                1 files):                                                                     │ │
│ │                C:\Users\mwebe\OneDrive\GitProjects\hysetter\data\example_1\streamcat_attrs\… │ │
│ │                │   │                                                                         │ │
│ │                nldi_attrs=FileList(C:\Users\mwebe\OneDrive\GitProjects\hysetter\data\exampl… │ │
│ │                0 files):                                                                     │ │
│ │                ,                                                                             │ │
│ │                │   │                                                                         │ │
│ │                forcing=FileList(C:\Users\mwebe\OneDrive\GitProjects\hysetter\data\example_1… │ │
│ │                1 files):                                                                     │ │
│ │                C:\Users\mwebe\OneDrive\GitProjects\hysetter\data\example_1\forcing\daymet_2… │ │
│ │                │   │                                                                         │ │
│ │                topo=FileList(C:\Users\mwebe\OneDrive\GitProjects\hysetter\data\example_1\to… │ │
│ │                1 files):                                                                     │ │
│ │                C:\Users\mwebe\OneDrive\GitProjects\hysetter\data\example_1\topo\topo_geom_0… │ │
│ │                │   │                                                                         │ │
│ │                soil=FileList(C:\Users\mwebe\OneDrive\GitProjects\hysetter\data\example_1\so… │ │
│ │                1 files):                                                                     │ │
│ │                C:\Users\mwebe\OneDrive\GitProjects\hysetter\data\example_1\soil\soilgrids_g… │ │
│ │                │   │                                                                         │ │
│ │                nlcd=FileList(C:\Users\mwebe\OneDrive\GitProjects\hysetter\data\example_1\nl… │ │
│ │                1 files):                                                                     │ │
│ │                C:\Users\mwebe\OneDrive\GitProjects\hysetter\data\example_1\nlcd\nlcd_geom_0… │ │
│ │                │   │                                                                         │ │
│ │                nid=FileList(C:\Users\mwebe\OneDrive\GitProjects\hysetter\data\example_1\nid, │ │
│ │                1 files):                                                                     │ │
│ │                C:\Users\mwebe\OneDrive\GitProjects\hysetter\data\example_1\nid\nid_geom_0.p… │ │
│ │                │   │                                                                         │ │
│ │                streamflow=FileList(C:\Users\mwebe\OneDrive\GitProjects\hysetter\data\exampl… │ │
│ │                1 files):                                                                     │ │
│ │                C:\Users\mwebe\OneDrive\GitProjects\hysetter\data\example_1\streamflow\strea… │ │
│ │                00:00:00_2018-01-02 00:00:00.nc,                                              │ │
│ │                │   │   remote_rasters={                                                      │ │
│ │                │   │   │   'rasters':                                                        │ │
│ │                FileList(C:\Users\mwebe\OneDrive\GitProjects\hysetter\data\example_1\rasters, │ │
│ │                0 files):                                                                     │ │
│ │                │   │                                                                         │ │
│ │                │   │   }                                                                     │ │
│ │                │   )                                                                         │ │
│ │                )                                                                             │ │
│ │         nwis = <pygeohydro.nwis.NWIS object at 0x00000204448DAF00>                           │ │
│ │         qobs = <xarray.Dataset> Size: 400B                                                   │ │
│ │                Dimensions:       (time: 2, station_id: 1)                                    │ │
│ │                Coordinates:                                                                  │ │
│ │                  * time          (time) datetime64[ns] 16B 2018-01-01T08:00:00               │ │
│ │                2018-01-02T08...                                                              │ │
[…]
