### 2025-06-19 ba39a41
fix: Fix a bug in the NWIS module where due to having : in filenames, Windows   users were unable to download streamflow data. The bug was fixed by removing   hours and minutes from the file name. #7 [skip ci]
