### 2025-06-19 dd08e0e
docs: In the installation section, explain that  and Version: 2.1.1
mamba [OPTIONS] [SUBCOMMAND]

OPTIONS:
  -h,     --help              Print this help message and exit
          --version

Configuration options:
          --rc-file FILE1 FILE2...
                              Paths to the configuration files to use
          --no-rc             Disable the use of configuration files
          --no-env            Disable the use of environment variables

Global options:
  -v,     --verbose           Set verbosity (higher verbosity with multiple -v, e.g. -vvv)
          --log-level ENUM:value 
[…]
