### 2025-06-19 7c3e54c
docs: update the changelog. [skip ci]
