+++ CONTRIBUTING.md
@@ -25,10 +25,10 @@ experience for all involved. The community looks forward to your contributions. 
 
 - [I Have a Question](#i-have-a-question)
 - [I Want To Contribute](#i-want-to-contribute)
-    - [Reporting Bugs](#reporting-bugs)
-    - [Suggesting Enhancements](#suggesting-enhancements)
-    - [Your First Code Contribution](#your-first-code-contribution)
-    - [Improving The Documentation](#improving-the-documentation)
+  - [Reporting Bugs](#reporting-bugs)
+  - [Suggesting Enhancements](#suggesting-enhancements)
+  - [Your First Code Contribution](#your-first-code-contribution)
+  - [Improving The Documentation](#improving-the-documentation)
 
 ## I Have a Question
 
@@ -46,7 +46,7 @@ the following:
 - Open an [Issue](https://github.com/hyriver/hysetter/issues/new).
 - Provide as much context as you can about what you're running into.
 - Provide project and platform versions (`python`, `pixi`, etc), depending on what seems
-    relevant.
+  relevant.
 
 We will then take care of the issue as soon as possible.
 
@@ -71,22 +71,22 @@ us fix any potential bug as fast as possible.
 
 - Make sure that you are using the latest version.
 - Determine if your bug is really a bug and not an error on your side e.g. using
-    incompatible environment components/versions (Make sure that you have read the
-    [documentation](https://hyriver.github.io/hysetter). If you are looking for support,
-    you might want to check [this section](#i-have-a-question)).
+  incompatible environment components/versions (Make sure that you have read the
+  [documentation](https://hyriver.github.io/hysetter). If you are looking for support,
+  you might want to check [this section](#i-have-a-question)).
 - To see if other users have experienced (and potentially already solved) the same issue
-    you are having, check if there is not already a bug report existing for your bug or
-    error in the [bug tracker](https://github.com/hyriver/hysetterissues?q=label%3Abug).
+  you are having, check if there is not already a bug report existing for your bug or
+  error in the [bug tracker](https://github.com/hyriver/hysetterissues?q=label%3Abug).
 - Also make sure to search the internet (including Stack Overflow) to see if users
-    outside of the GitHub community have discussed the issue.
+  outside of the GitHub community have discussed the issue.
 - Collect information about the bug:
-    - Stack trace (Traceback)
-    - OS, Platform and Version (Windows, Linux, macOS, x86, ARM)
-    - Version of the interpreter, compiler, SDK, runtime environment, package manager,
-        depending on what seems relevant.
-    - Possibly your input and the output
-    - Can you reliably reproduce the issue? And can you also reproduce it with older
-        versions?
+  - Stack trace (Traceback)
+  - OS, Platform and Version (Windows, Linux, macOS, x86, ARM)
+  - Version of the interpreter, compiler, SDK, runtime environment, package manager,
+    depending on what seems relevant.
+  - Possibly your input and the output
+  - Can you reliably reproduce the issue? And can you also reproduce it with older
+    versions?
 
 <!-- omit in toc -->
 
@@ -100,25 +100,25 @@ We use GitHub issues to track bugs and errors. If you run into an issue with the
 project:
 
 - Open an [Issue](https://github.com/hyriver/hysetter/issues/new). (Since we can't be
-    sure at this point whether it is a bug or not, we ask you not to talk about a bug
-    yet and not to label the issue.)
+  sure at this point whether it is a bug or not, we ask you not to talk about a bug yet
+  and not to label the issue.)
 - Explain the behavior you would expect and the actual behavior.
 - Please provide as much context as possible and describe the *reproduction steps* that
-    someone else can follow to recreate the issue on their own. This usually includes
[…]
