+++ docs/examples/index.md
@@ -18,6 +18,6 @@ environment, so you can select it as the interpreter for the notebooks.
 
 - [![HYMOD](images/hymod-flowchart.png){ loading=lazy }](hymod.ipynb "HYMOD") **HYMOD**
 - [![NCRS-PDM](images/ncrspdm.png){ loading=lazy }](ncrspdm.ipynb "NCRS-PDM")
-    **NCRS-PDM**
+  **NCRS-PDM**
 
 </div>

[…]
