### 2025-06-19 8b1a187
chore: bump pixi version. [skip ci]
