+++ .github/workflows/test.yml
@@ -26,9 +26,9 @@ jobs:
 
     steps:
     - uses: actions/checkout@v4
-    - uses: prefix-dev/setup-pixi@v0.8.8
+    - uses: prefix-dev/setup-pixi@v0.8.9
       with:
-        pixi-version: v0.47.0
+        pixi-version: v0.48.2
         environments: ${{ matrix.environment }}
         activate-environment: ${{ matrix.environment }}
         cache: true
