### 2025-06-19 2728e86
chore: update pre-commit hooks. [skip ci]
