+++ pyproject.toml
@@ -186,6 +186,8 @@ lint.ignore = [
   "ISC001",
   # specific type error ignored
   "PGH003",
+  # top level imports
+  "PLC0415",
   "PLR0913",
   "PLR2004",
   # url schema
