### 2025-06-19 89f1797
docs: add instructions for using mamba or conda instead of micromamba. #6 [skip ci]
