### 2025-06-20 hyriver/hysetter#7 @mhweber
@cheginit that did the trick!  All worked with the config_demo.yml file for me after installing the latest from `git`.
