### 2025-06-24 8b191a0
docs: add installation instruction for both pixi and conda. #8 [skip ci]
