### 2025-06-24 670905d
chore: run pre-commit. [skip ci]
