### 2025-06-24 8f87aa6
chore: exclude the paper from mdformat since it messes up the paper title. [skip ci]
