+++ CHANGELOG.md
@@ -10,14 +10,14 @@ this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.htm
 ### Fixed
 
 - Fix a bug in the NWIS module where due to having ":" in filenames, Windows users were
-  unable to download streamflow data. The bug was fixed by removing hours and minutes
-  from the file name. [#7](https://github.com/hyriver/hysetter/issues/7)
+    unable to download streamflow data. The bug was fixed by removing hours and minutes
+    from the file name. [#7](https://github.com/hyriver/hysetter/issues/7)
 
 ### Added
 
 - In the installation section, explain that `conda` and `mamba` can be used instead of
-  `micromamba` for installing HySetter.
-  [#6](https://github.com/hyriver/hysetter/issues/6)
+    `micromamba` for installing HySetter.
+    [#6](https://github.com/hyriver/hysetter/issues/6)
 
 ### Changed
 
@@ -26,13 +26,13 @@ this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.htm
 ### Fixed
 
 - Fix a bug in CLI where passing `--overwrite` removed the root directory instead of the
-  project directory. As a result, the root directory for the AOI file was removed
-  causing the CLI to fail.
+    project directory. As a result, the root directory for the AOI file was removed
+    causing the CLI to fail.
 
 ### Changed
 
 - For forcing and streamflow data, include the start and end dates in the file name to
-  avoid overwriting the files when the same data is requested for different dates.
+    avoid overwriting the files when the same data is requested for different dates.
 
 ## [0.3.0] - 2025-02-11
 
@@ -43,24 +43,25 @@ development. The most notable changes are listed below.
 
 - Add support for subsetting remote raster datasets.
 - Add support for POLARIS data which is a probabilistic remapping of SSURGO (POLARIS)
-  soil properties at 30-m resolution over the CONUS.
+    soil properties at 30-m resolution over the CONUS.
 - The `Config.file_paths` attribute now includes a `FileList` object for each requested
-  data source. This allows for easy access to the file paths of the downloaded data. The
-  indices of the list of files correspond to the geometries in the AOI GeoDataFrame.
+    data source. This allows for easy access to the file paths of the downloaded data.
+    The indices of the list of files correspond to the geometries in the AOI
+    GeoDataFrame.
 - Add a new option called `crop` to all gridded data sources. This option allows to
-  disable cropping of the data to the AOI. This is useful when the user wants to get the
-  data for the bounding box of the AOI instead of the AOI itself, e.g., for topography
-  data where the data at the edges of the AOI is needed.
+    disable cropping of the data to the AOI. This is useful when the user wants to get
+    the data for the bounding box of the AOI instead of the AOI itself, e.g., for
+    topography data where the data at the edges of the AOI is needed.
 - Add a new option called `geometry_buffer` to all gridded data sources. This option
-  allows to buffer the geometry of the AOI by a given distance. This is useful when the
-  user wants to get the data for a larger area than the AOI itself, e.g., for topography
-  data where the data at the edges of the AOI is needed.
+    allows to buffer the geometry of the AOI by a given distance. This is useful when
+    the user wants to get the data for a larger area than the AOI itself, e.g., for
+    topography data where the data at the edges of the AOI is needed.
 - Add a new option called `overwrite` to allow overwriting the existing data files.
 
 ### Changed
 
 - A major refactoring of the code base to improve the performance by avoiding
-  unnecessary web service calls and package imports.
+    unnecessary web service calls and package imports.
 
 ## [0.2.0] - 2024-09-13
 
[…]
