+++ CODE_OF_CONDUCT.md
@@ -20,19 +20,19 @@ include:
 - Being respectful of differing opinions, viewpoints, and experiences
 - Giving and gracefully accepting constructive feedback
 - Accepting responsibility and apologizing to those affected by our mistakes, and
-  learning from the experience
+    learning from the experience
 - Focusing on what is best not just for us as individuals, but for the overall community
 
 Examples of unacceptable behavior include:
 
 - The use of sexualized language or imagery, and sexual attention or advances of any
-  kind
+    kind
 - Trolling, insulting or derogatory comments, and personal or political attacks
 - Public or private harassment
 - Publishing others' private information, such as a physical or email address, without
-  their explicit permission
+    their explicit permission
 - Other conduct which could reasonably be considered inappropriate in a professional
-  setting
+    setting
 
 ## Enforcement Responsibilities
 
