### 2025-06-24 869af72
docs: fix the env generation command. [skip ci]
