### 2025-06-24 7ce8e23
chore: remove the gdal pin. [skip ci]
