### 2025-06-24 b7b722c
docs: use attributes instead of parameters for class docstrings. [skip ci]
