### 2025-06-25 hyriver/hysetter#7 @mhweber
context reference [openjournals/joss-reviews#8301](https://github.com/openjournals/joss-reviews/issues/8301)
