### 2025-07-08 c2ddf27
JOSS Edits (#1)
* disable markdown formatter, render LICENSE easier to be machine-read

* small grammar fixes, some internal variable renaming

* autoformatting

* make note about autoref links

* Update src/hysetter/hysetter.py

* include license in docs

* small formatting changes
