+++ LICENSE.md
@@ -1,7 +1,3 @@
-=======
-License
-=======
-
 MIT License
 
 Copyright (c) 2024, Taher Chegini
