### 2025-07-10 #8301 @Zeitsperre
@cheginit Absolutely, will do! If all goes according to plan, I should be done tomorrow morning (EST).

There is one thing I added in my PR: https://github.com/hyriver/hysetter/commit/c07bd8e3bd53e8cb8a029cc9f95b2c734b03c2a2#diff-a0198f95473849eaebe4f3b238ede8ea1049d03f0ec252df8cf7f25845eca217R65

I don't think that particular LaTeX formatting was working in the paper. Other than that, I don't anticipate too much else yet.
