### 2025-07-14 hyriver/hysetter#11 @Zeitsperre
[JOSS] Autoref links are broken in draft paper
### What happened?

It seems that `pandoc` does not support the `autoref` LaTeX command. Inline references to the figures need to be adjusted in the `paper.md` to use `pandoc` commands instead.

### Potential Solution (untested)

```Markdown
@fig:config-example shows a typical configuration file, and @fig:config-demo demonstrates
HySetter's command-line interface with progress indicators.

![Example YAML configuration file for HySetter specifying project parameters, area of interest, and data sources.abel{fig:config-example}](config.png){#fig:config-example width="300pt"}

![HySetter command-line interface showing real-time progress during data acquisition processes.abel{fig:config-demo}](cli.png){#fig:config-demo}

HySetter currently supports subsetting from the following data sources:

- **Area of Interest**: HUC watersheds, GAGES-II basins, NHDPlusV2 catchments and their
    attributes (StreamCat and NLDI), or user-defined geometries
- **Drainage Network**: NHDPlusV2
- **Climate Forcing**: Daymet, GridMET, NLDAS2
- **Streamflow**: National Water Information System (NWIS)
- **Soil**: gNATSGO, SoilGrids
- **Topography**: 3D Elevation Program (3DEP)
- **Dams**: National Inventory of Dams (NID)
- **Land Use/Land Cover**: Multi-Resolution Land Characteristics (MRLC) products
    including land cover, canopy, imperviousness, and urban descriptors

![HySetter workflow showing data flow from remote dataset repositories and web services through HyRiver to create area-specific datasets for hydrological modeling.abel{fig:flowchart}](flowchart.png){#fig:flowchart width="350pt"}

...
```

### Anything else we need to know?

This issue needs to be addressed for JOSS review approval.

Cross-reference: https://github.com/openjournals/joss-reviews/issues/8301

### Environment

N/A
