### 2025-07-14 d93f17f
docs: fix the issue with the icon link. (#10) [skip ci]
