### 2025-07-17 #8301 @jGaboardi
Nice work on this @cheginit - now this is over to the track editor for further processing.

And thank you @mhweber  & @Zeitsperre  for your timely and thoughtful reviews!
