### 2025-07-17 #8301 @editorialbot
:wave: @openjournals/ese-eics, this paper is ready to be accepted and published.
