### 2025-07-17 2f0424b
chore: fix the license issue for the conda-forge release. [skip ci]
