### 2025-07-17 #8301 @jGaboardi
@Zeitsperre – Good catch. @cheginit Did you create that tag+release after you sent the DOI earlier today? It seems [that was tagged](https://github.com/hyriver/hysetter/releases/tag/v0.3.4) after I [recommended accept](https://github.com/openjournals/joss-reviews/issues/8301#issuecomment-3084757993.)
