### 2025-07-17 #8301 @cheginit
Since the license file was renamed during the revisions, the `conda-forge` build failed. So I just fixed. There's no need to change the version and DOI, unless you think it's necessary.
