### 2025-07-18 #8301 @kthyng
Thanks @cheginit I think we can move forward as is.
