### 2025-07-18 #8301 @jGaboardi
> Since the license file was renamed during the revisions, the `conda-forge` build failed. So I just fixed. There's no need to change the version and DOI, unless you think it's necessary.

@cheginit Yep, that's no problem at all. I just wanted to check to make sure I didn't miss something.
